# public/openapi-v1.yaml
openapi: 3.1.0
info:
  title: Syntara Public API
  version: 1.0.0
  description: Public API for the first batch of Syntara learning workflows.
servers:
  - url: /
security:
  - bearerAuth: []
paths:
...
  /api/v1/courses/{courseId}/questions:
    post:
      operationId: createCourseQuestion
      description: >-
        Answers one course-scoped question from server-trusted course sources,
        problem evidence, and layered learner memory. The completed turn is
        persisted in the same conversation store used by /learn. Memory write
        actions remain pending confirmation.
      parameters:
        - in: path
          name: courseId
          required: true
          schema: { type: string }
        - in: header
          name: Idempotency-Key
          required: true
          schema: { type: string, minLength: 1, maxLength: 200 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CourseQuestionRequest' }
      responses:
        '201':
          description: Newly generated and persisted course answer
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessEnvelope' }
        '200':
          description: Idempotent replay of the original persisted answer
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessEnvelope' }
        '400':
          description: Invalid request or missing Idempotency-Key
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '404':
          description: Course not found for the authenticated API principal
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '409':
          description: Idempotency conflict or an identical request is still in progress
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Syntara API Key
  responses:
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
    SuccessEnvelope:
      type: object
      required: [success, request_id, data]
      properties:
        success: { type: boolean, const: true }
        request_id: { type: string }
        data: { type: object, additionalProperties: true }
    ErrorEnvelope:
      type: object
      required: [success, request_id, error]
      properties:
        success: { type: boolean, const: false }
        request_id: { type: string }
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: {}
...
    CourseQuestionRequest:
      type: object
      required: [question]
      additionalProperties: false
      properties:
        question: { type: string, minLength: 1, maxLength: 12000 }
        session_id:
          type: string
          minLength: 1
          maxLength: 160
          pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]*$'
        options:
          type: object
          additionalProperties: false
          properties:
            include_evidence: { type: boolean, default: true }
            include_memory_write_proposals: { type: boolean, default: true }